start looking at removing specflow internal
