+++ src/NUnitRetry.ReqnrollPlugin/NUnitRetry.ReqnrollPlugin.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<TargetFrameworks>netstandard2.0</TargetFrameworks>
@@ -8,7 +8,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Reqnroll.CustomPlugin" Version="2.3.0" />
-		<PackageReference Include="SpecFlow.Internal.Json" Version="1.0.8" />
+		<PackageReference Include="Reqnroll.CustomPlugin" Version="2.4.0" />
 	</ItemGroup>
 </Project>
