+++ src/NUnitRetry.ReqnrollPlugin/NUnitRetry.ReqnrollPlugin.nuspec
@@ -1,18 +1,17 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
   <metadata>
     <id>NUnitRetry.ReqnrollPlugin</id>
-    <version>1.0.300</version>
+    <version>1.1.0-preview-3</version>
     <authors>Josh Keegan, Piotr Niedzialek, Chris Billson</authors>
     <owners>Chris Billson</owners>
     <projectUrl>https://github.com/chrisbillson/NUnitRetry.ReqnrollPlugin</projectUrl>
     <license type="expression">MIT</license>
     <description>Allows for a @Retry, @Retry(n) tags to be placed above Scenarios and Features in Reqnroll.NUnit. This can be used to re-run a failed test up to n times, allowing for transient failures (e.g. network related) to be tolerated. Based on Josh Keegan's xRetry.SpecFlow package.</description>
-    <copyright>2022</copyright>
+    <copyright>2025</copyright>
     <tags>Reqnroll NUnit Retry Test-Automation</tags>
 
     <dependencies>
-	    <dependency id="Reqnroll.CustomPlugin" version="2.3.0" />
-	    <dependency id="SpecFlow.Internal.Json" version="1.0.8" />
+	    <dependency id="Reqnroll.CustomPlugin" version="2.4.0" />
     </dependencies>
   </metadata>
