+++ src/NUnitRetry.ReqnrollPlugin/TestGeneratorProvider.cs
@@ -5,6 +5,7 @@ using System.CodeDom;
 using System.Collections.Generic;
 using System.Linq;
 using Reqnroll.Generator;
+using NUnitRetry.ReqnrollPlugin.Configuration;
 
 namespace NUnitRetry.ReqnrollPlugin
 {
@@ -12,9 +13,9 @@ namespace NUnitRetry.ReqnrollPlugin
     {
         protected internal const string RETRY_ATTR = "NUnit.Framework.Retry";
 
-        private readonly Configuration _configuration;
+        private readonly RetryConfiguration _configuration;
 
-        public TestGeneratorProvider(CodeDomHelper codeDomHelper, Configuration configuration) : base(codeDomHelper)
+        public TestGeneratorProvider(CodeDomHelper codeDomHelper, RetryConfiguration configuration) : base(codeDomHelper)
         {
             _configuration = configuration;
         }
