+++ tests/NUnitRetry.Tests/NUnitRetry.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<TargetFramework>net8</TargetFramework>
@@ -9,8 +9,8 @@
 		<PackageReference Include="FluentAssertions" Version="5.10.3" />
         <PackageReference Include="NUnit" Version="3.13.2" />
         <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
-        <PackageReference Include="NUnitRetry.ReqnrollPlugin" Version="1.0.300" />
-        <PackageReference Include="Reqnroll.NUnit" Version="2.3.0" />
+        <PackageReference Include="NUnitRetry.ReqnrollPlugin" Version="1.1.0-preview-3" />
+        <PackageReference Include="Reqnroll.NUnit" Version="2.4.0" />
 	</ItemGroup>
 
 	<ItemGroup>
